+++ pkg/plugininventory/sqlite_inventory_metadata_test.go
